+++ package.json
@@ -21,6 +21,7 @@
     "gray-matter": "^4.0.3",
     "markdown-it": "^13.0.1",
     "next": "^12.0.10",
+    "next-language-detector": "^1.0.2",
     "next-mdx-remote": "^4.0.2",
     "next-seo": "^5.1.0",
     "react": "^17.0.2",
