+++ src/lang/en.json
@@ -1,4 +1,6 @@
 {
+  "page.index.meta.title": "Featured work by Pablo Romero Jarén. UI/Graphic Artist",
+  "page.index.meta.description": "UI/Graphic Artist",
   "page.index.title.work": "UI/Graphic Artist",
   "page.index.title.whoami": "Who am I?",
   "page.index.title.contact": "Thanks for visiting!",
