+++ src/lang/es.json
@@ -1,4 +1,6 @@
 {
+  "page.index.meta.title": "Trabajos destacados de Pablo Romero Jarén. Artista UI/Gráfico",
+  "page.index.meta.description": "Artista UI/Gráfico",
   "page.index.title.work": "Artista UI/Gráfico",
   "page.index.title.whoami": "¿Quien soy?",
   "page.index.title.contact": "¡Gracias por la visita!",
