+++ src/pages/_document.tsx
@@ -1,13 +1,16 @@
 /* eslint-disable @next/next/no-sync-scripts */
 import Document, { Html, Head, Main, NextScript } from 'next/document';
 
-import { AppConfig } from '../utils/AppConfig';
+import { DEFAULT_LOCALE } from '../utils/lang';
 
 // Need to create a custom _document because i18n support is not compatible with `next export`.
 class MyDocument extends Document {
   render() {
+    const currentLocale =
+      // eslint-disable-next-line no-underscore-dangle
+      this.props.__NEXT_DATA__.query.locale || DEFAULT_LOCALE;
     return (
-      <Html lang={AppConfig.locale}>
+      <Html lang={currentLocale as string}>
         <Head />
         <body>
           <Main />
