+++ src/utils/AppConfig.ts
@@ -5,5 +5,5 @@ export const AppConfig = {
   url: 'https://example.com',
   locale: 'en',
   author: 'Pablo Romero Jarén',
-  pagination_size: 20,
+  pagination_size: 100,
 };
