+++ src/utils/lang.ts
@@ -1,7 +1,16 @@
+import nextLanguageDetector from 'next-language-detector';
+
 import en from '../lang/en.json';
 import es from '../lang/es.json';
 
+export type SupportedLocales = 'en' | 'es';
+
 export const DEFAULT_LOCALE = 'en';
+
+export const locales: { [k: string]: SupportedLocales } = {
+  en: 'en',
+  es: 'es',
+};
 
 export const localeMessages = {
   es,
@@ -13,4 +22,7 @@ export type LocaleMessages = {
   es: typeof es;
 };
 
-export type SupportedLocales = 'en' | 'es';
+export const languageDetector = nextLanguageDetector({
+  supportedLngs: Object.keys(locales),
+  fallbackLng: DEFAULT_LOCALE,
+});
